ddb/requirements.txt
pygdbmi # gdbmi parser
pyyaml # yaml parser
paramiko # ssh command execution
kubernetes
paho-mqtt # mqtt client